% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_combs_ensemble.R
\name{all_combs_ensemble}
\alias{all_combs_ensemble}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
all_combs_ensemble(
  data,
  model_dist_data,
  avail_threshold,
  nmod,
  window = 5,
  init_weeks = 5,
  excl = c("EuroCOVIDhub-baseline", "EuroCOVIDhub-ensemble")
)
}
\arguments{
\item{data}{data (subset or full) from the European Forecast hub}

\item{model_dist_data}{model distance data (output from model_dist_by_fc_date)}

\item{avail_threshold}{minimum availability for models to be considered in the
sampling process}

\item{nmod}{number of models to sample at each date}

\item{window}{window size for recent distance computation}

\item{init_weeks}{number of weeks left out at beginning for init}

\item{excl}{which models should be excluded from the ensemble experiment}
}
\description{
Samples ALL combinations of possible ensembles of size n at given timepoint,
scores them and computes average pairwise historical and recent distance
}
