% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{basic_dist_ovr_hor}
\alias{basic_dist_ovr_hor}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
basic_dist_ovr_hor(plot_dat, quantiles = seq(0.1, 0.9, 0.1))
}
\description{
Plots model_distance vs performance
}
