% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_experiments.R
\name{best_performers_ensemble}
\alias{best_performers_ensemble}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
best_performers_ensemble(
  data,
  score = "interval_score",
  nmods = 3,
  window = 4,
  may_miss = 0,
  excl = c("EuroCOVIDhub-baseline", "EuroCOVIDhub-ensemble"),
  cvg_threshold = NULL,
  return_model_data = FALSE
)
}
\arguments{
\item{data}{data (subset or full) from the European Forecast hub}

\item{score}{which scoring function should be used to select best models}

\item{nmods}{number of models to sample}

\item{window}{number of weeks to consider in establishing what
the best models are}

\item{may_miss}{number of forecasts within the rolling window a model
may miss to be eligible for the best performers ensemble}

\item{excl}{models to exclude from the ensemble}

\item{cvg_threshold}{minimum coverage required for models to be
included in the sampling process}

\item{return_model_data}{if set to TRUE, additionally to the ensemble predictions,
a list of matrices indicating which models were used in which step
are returned}
}
\value{
by default, just the ensemble predictions. If \code{return_model_data}
is set to TRUE, returns a list with the ensemble predictions as well
as a list of matrices that indicate which model was used in each round
of the best performers ensemble
}
\description{
Assesses ensemble performance by iteratively sampling given numbers of
model for the (mean/median) ensemble

Uses fun_make_ensemble.R
}
