% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ensemble_assessment.R
\name{compute_decomp_scores}
\alias{compute_decomp_scores}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
compute_decomp_scores(data, adjust_avail = TRUE)
}
\arguments{
\item{adjust_avail}{should each model or each model - forecast date - location instance
be weighted equally (default: TRUE, i.e. equal weight of each model - forecast date - location
instance; this prevents models that have low availability to influence average scores too much)}

\item{hub_data}{data from the European forecast hub}

\item{coverage_ranges}{PI ranges to compute coverage for}

\item{avail_threshold}{threshold of availability for models to be included}
}
\value{
data.table (long format) with decomposed WIS at level of model types
}
\description{
Helper function for decomposition of WIS for model types
}
