% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cramers_dist}
\alias{cramers_dist}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
cramers_dist(q_F, q_G, tau)
}
\arguments{
\item{q_F}{quantiles of distribution F}

\item{q_G}{quantiles of distribution G}

\item{tau}{quantile levels of F and G}
}
\value{
Cramer distance ()
}
\description{
Helper function to compute the Cramer distance.
Minimally adapted from: https://github.com/reichlab/covidHubUtils/blob/master/R/calc_cramers_dist_unequal_space.R
}
