% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fc_dates_window}
\alias{fc_dates_window}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
fc_dates_window(curr_date, window, incl = FALSE)
}
\arguments{
\item{curr_date}{the current forecast date}

\item{window}{size of the sliding window}

\item{incl}{should current date be in included in sliding window (default is
FALSE, which should be used if current date's information has not realized yet.
for e.g. distance calculations between forecasts, can be set to TRUE)}
}
\value{
vector with dates in sliding window
}
\description{
Helper function to give forecast dates in sliding window
}
