% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getmodels}
\alias{getmodels}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
getmodels(
  data,
  excl = c("EuroCOVIDhub-baseline", "EuroCOVIDhub-ensemble"),
  incl = NULL,
  avail_threshold = NULL
)
}
\arguments{
\item{data}{data (subset or full) from the European Forecast hub}

\item{excl}{vector of models to be excluded}

\item{incl}{vector of models to be included}

\item{avail_threshold}{threshold of availability for model to be included}
}
\value{
either a data.frame with only the models and their respective
availability, or the original data.table including and extra column
with the availability values of the respective models (the default)
}
\description{
Helper function that returns models given excl and an availability threshold
Also checks whether there still are ensembles in the data (this is because
this function is typically used with an ensembling procedure, where ensemble
models are not wanted as input)
}
