% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{make_ensemble}
\alias{make_ensemble}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
make_ensemble(
  data,
  summary_function = median,
  model_name = NULL,
  excl = c("EuroCOVIDhub-baseline", "EuroCOVIDhub-ensemble"),
  extra_excl = NULL,
  incl = NULL,
  strat = c("location", "forecast_date", "quantile", "horizon", "target_type"),
  extra_vars = c("target_end_date", "n", "population"),
  avail_threshold = NULL
)
}
\arguments{
\item{data}{data (subset or full) from the European Forecast hub}

\item{summary_function}{function to build ensemble from (median by default)}

\item{model_name}{name for the resulting model. By default, makes a name based
on summary_function}

\item{excl}{models to exclude from ensemble (by default the official ensemble
and the baseline)}

\item{extra_excl}{extra models to exclude, apart from excl}

\item{incl}{explicit list of models that should be included in the ensemble}

\item{strat}{how to stratify, i.e. unit defining a single forecast}

\item{extra_vars}{any extra (redundant) variables that should stay in the data}

\item{avail_threshold}{minimum availability for models to be included in
the ensemble}
}
\value{
a data.table object that contains the ensemble forecasts in addition to the
models that are already in data
}
\description{
Ensemble builder function for data from the European Forecast Hub
By default, builds ensemble based on all models except the baseline and Hub
ensemble models. More models to exclude or, for the reverse approach, a list
of models to include may be supplied
}
