% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{make_ensemble_minimal}
\alias{make_ensemble_minimal}
\title{@title COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
make_ensemble_minimal(
  data,
  which_ens = NULL,
  excl = c("EuroCOVIDhub-baseline", "EuroCOVIDhub-ensemble"),
  extra_excl = NULL,
  strat = c("location", "forecast_date", "quantile", "horizon", "target_type")
)
}
\arguments{
\item{data}{data (subset or full) from the European Forecast hub}

\item{excl}{models to exclude from ensemble (by default the official ensemble
and the baseline)}

\item{extra_excl}{extra models to exclude, apart from excl}

\item{strat}{how to stratify, i.e. unit defining a single forecast}

\item{summary_function}{function to build ensemble from (median by default)}

\item{model_name}{name for the resulting model. By default, makes a name based
on summary_function}

\item{incl}{explicit list of models that should be included in the ensemble}

\item{extra_vars}{any extra (redundant) variables that should stay in the data}

\item{avail_threshold}{minimum availability for models to be included in
the ensemble}
}
\value{
a data.table object that contains the ensemble forecasts
}
\description{
Minimal Ensemble builder function for data from the European Forecast Hub
Used with make_ensemble, has some features removed. Not recommended for other uses.
Different from other version: takes reduced dataset, only returns ensemble data
and not other models
}
