% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ensemble_assessment.R
\name{make_eval}
\alias{make_eval}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
make_eval(
  target,
  current,
  su_cols,
  scoring_fun = "interval_score",
  strat_by = c("model", "target_type", "location", "horizon", "forecast_date"),
  return_anti_join = FALSE
)
}
\arguments{
\item{target}{ensemble data to evaluate score against current}

\item{current}{comparison data}

\item{su_cols}{columns for scoringutils}

\item{scoring_fun}{which scoring function to evaluate}

\item{strat_by}{unit of a single forecast}

\item{return_anti_join}{if target's and current's targets don't match (function
gives and error), should setdiff between the two be returned}
}
\value{
data.table (long format) with relative scores
}
\description{
Helper function to evaluate different ensemble strategies
}
