% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model_coverage}
\alias{model_coverage}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
model_coverage(data, alldat = TRUE)
}
\arguments{
\item{data}{data (subset or full) from the European Forecast hub}

\item{alldat}{should full data be returned with extra column for
coverage added (the default) or only a dataframe with
the coverage data (columns are models, location,
target_type)}
}
\value{
either a data.frame with only the models and their respective
coverage, or the original data.table including and extra column
with the coverage values of the respective models (the default)
}
\description{
Computes coverage per model and location over entire observation time
}
