% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model_dist}
\alias{model_dist}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
model_dist(
  data,
  avail_threshold,
  avail_overlap_threshold,
  dist_fun,
  excl = c("EuroCOVIDhub-baseline", "EuroCOVIDhub-ensemble"),
  normalize = FALSE,
  return_symmetric = TRUE
)
}
\arguments{
\item{data}{data (subset or full) from the European Forecast hub}

\item{avail_threshold}{minimum availability for a model to be considered}

\item{excl}{models to exclude from the calculation}
}
\value{
either a data.frame with only the models and their respective
availability, or the original data.table including and extra column
with the availability values of the respective models (the default)
}
\description{
Helper function that computes a distance measure for each pair of models
in the European Forecast Hub (a forecast in the Hub is distributional
and is comprised of a discrete set of 23 forecast quantiles)
}
