% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model_dist_by_fc_date}
\alias{model_dist_by_fc_date}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
model_dist_by_fc_date(
  data,
  avail_threshold,
  avail_overlap_threshold,
  dist_fun,
  excl = c("EuroCOVIDhub-baseline", "EuroCOVIDhub-ensemble")
)
}
\arguments{
\item{data}{data (subset or full) from the European Forecast hub}

\item{avail_threshold}{minimum availability for a model to be considered}

\item{avail_overlap_threshold}{minimum availability overlap for a model pair
to have their distance computed}

\item{dist_fun}{which distance function to use}

\item{excl}{models to exclude from the calculation}
}
\value{
either a data.frame with only the models and their respective
availability, or the original data.table including and extra column
with the availability values of the respective models (the default)
}
\description{
Helper function that computes a distance measure for each pair of models
in the European Forecast Hub (a forecast in the Hub is distributional
and is comprised of a discrete set of 23 forecast quantiles)
}
