% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model_dists_to_mat}
\alias{model_dists_to_mat}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
model_dists_to_mat(model_dist_dt)
}
\arguments{
\item{model_dist_mat}{single matrix from model_dist output (list)}
}
\value{
matrix as data.table
}
\description{
Helper function to transform output from model_dist
}
