% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ensemble_assessment.R
\name{overall_assessment_model_types}
\alias{overall_assessment_model_types}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
overall_assessment_model_types(
  hub_data,
  coverage_ranges = c(50, 90),
  avail_threshold = 0,
  adjust_avail = TRUE
)
}
\arguments{
\item{hub_data}{data from the European forecast hub}

\item{coverage_ranges}{PI ranges to compute coverage for}

\item{avail_threshold}{threshold of availability for models to be included}

\item{adjust_avail}{should each model or each forecast date - location be weighted
equally (default: TRUE, i.e. forecast date - location)}
}
\value{
data.table
}
\description{
Helper function for assessment of model types
Includes coverage, bias
}
