% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{overall_assessment_plot}
\alias{overall_assessment_plot}
\title{Overall assessment plot}
\usage{
overall_assessment_plot(
  overall_score_data,
  decomp_data,
  relative = FALSE,
  breaks = c("mechanistic", "semi", "statistical"),
  labels = c("mech.", "semi", "stat.")
)
}
\description{
Overall assessment plot
}
