% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_best_performers_models}
\alias{plot_best_performers_models}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
plot_best_performers_models(
  best_performers_models,
  data,
  saveplot = TRUE,
  path = here("plots", "best_performers_models.pdf")
)
}
\arguments{
\item{best_performers_models}{model picking results from running
best_performers_ensemble}

\item{data}{full data (or subset of) from European Forecast Hub}

\item{saveplot}{should plotting output be saved}

\item{path}{where plotting output should be saved}
}
\description{
Plots model pick results of best_performers
}
