% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_best_performers_scores}
\alias{plot_best_performers_scores}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
plot_best_performers_scores(
  best_performers_scores,
  saveplot = TRUE,
  path = here("plots", "best_performers_scores.pdf")
)
}
\arguments{
\item{best_performers_scores}{score results from running
best_performers_ensemble}

\item{saveplot}{should plotting output be saved}

\item{path}{where plotting output should be saved}
}
\description{
Plots score results of best performers
}
