% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_dist_vs_scores_overtime}
\alias{plot_dist_vs_scores_overtime}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
plot_dist_vs_scores_overtime(
  plot_data,
  score_data = NULL,
  palette = "Set2",
  min_x.ticks = TRUE
)
}
\description{
Plots model_distance vs performance
}
