% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_kickout_results}
\alias{plot_kickout_results}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
plot_kickout_results(
  model_kickout_results,
  random_kickout_results,
  plot_score = "interval_score",
  saveplot = TRUE,
  path = here("plots", "model_similarity_kickout.pdf"),
  plotsize = c(width = 9, height = 12)
)
}
\arguments{
\item{saveplot}{should plot output be saved}

\item{path}{where plot should be saves}

\item{model_dists}{output from model_similarity}

\item{which_combs}{which combinations of location and target_type
should be plotted (required format: GB.Deaths, DE.Cases,...)}
}
\description{
Plots the results of model_dist as heatmaps
}
