% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_model_availability}
\alias{plot_model_availability}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
plot_model_availability(
  data,
  saveplot = TRUE,
  path = (here("plots", "availability.pdf")),
  height = 12,
  width = 8,
  indiv = TRUE,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  palette = "Set1"
)
}
\arguments{
\item{data}{full data (or subset of) from European Forecast Hub}

\item{saveplot}{should plotting output be saved}

\item{path}{where plotting output should be saved}
}
\description{
Plots model availability
}
