% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_model_similarity}
\alias{plot_model_similarity}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
plot_model_similarity(
  model_dists,
  which_combs = NULL,
  saveplot = TRUE,
  path = here("plots", "model_similarity.pdf"),
  plotsize = c(width = 12.5, height = 9)
)
}
\arguments{
\item{model_dists}{output from model_similarity}

\item{which_combs}{which combinations of location and target_type
should be plotted (required format: GB.Deaths, DE.Cases,...)}

\item{saveplot}{should plot output be saved}

\item{path}{where plot should be saves}
}
\description{
Plots the results of model_dist as heatmaps
}
