% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_model_type_across_time}
\alias{plot_model_type_across_time}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
plot_model_type_across_time(pw_comp_dat)
}
\arguments{
\item{pw_comp_dat}{pairwise comparison data to plot}
}
\description{
Plots model type performance over periods and horizons
Code adapted from scoringutils package
}
\references{
Bosse NI, Gruson H, Cori A, van Leeuwen E, Funk S, Abbott S (2022).
“Evaluating Forecasts with scoringutils in R.”
arXiv. doi:10.48550/ARXIV.2205.07090, https://arxiv.org/abs/2205.07090.
}
