% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_trajectories}
\alias{plot_trajectories}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
plot_trajectories(
  data,
  saveplot = TRUE,
  path = here("plots", "trajectories.pdf")
)
}
\arguments{
\item{data}{full data (or subset of) from European Forecast Hub}

\item{saveplot}{should plotting output be saved}

\item{path}{where plotting output should be saved}
}
\description{
Plots trajectories of Deaths and Cases
}
