% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ensemble_assessment.R
\name{pw_comp_by_model_type}
\alias{pw_comp_by_model_type}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
pw_comp_by_model_type(
  data,
  su_cols,
  res_before_pw_comp = c("model_type", "forecast_date", "quantile", "location",
    "target_type", "horizon"),
  end_res = c("model", "horizon", "target_type")
)
}
\arguments{
\item{data}{data from the European forecast hub}

\item{su_cols}{cols defining unit of single forecast}

\item{res_before_pw_comp}{at what level to average across model_types
It is important that forecast_date, location are included here, otherwise
results will be skewed}

\item{end_res}{what level of stratification end result should have
default: model_type (it's only called more for scoringutils reasons), horizon, target_type}
}
\value{
data.table
}
\description{
Helper function for pairwise comparison of model types
}
