% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wasserstein_dist}
\alias{wasserstein_dist}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
wasserstein_dist(q_F, q_G, tau)
}
\arguments{
\item{q_F}{quantiles of distribution F}

\item{q_G}{quantiles of distribution G}

\item{tau}{quantile levels of F and G}
}
\value{
Wasserstein distance
}
\description{
Helper function to compute the Wasserstein 2 metric
for discrete quantiles
}
