% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{wis_decomp_model_type}
\alias{wis_decomp_model_type}
\title{WIS decomp}
\usage{
wis_decomp_model_type(
  score_data,
  relative = FALSE,
  breaks = c("mechanistic", "semi", "statistical"),
  labels = c("mech.", "semi", "stat.")
)
}
\value{
A ggplot2 theme
}
\description{
WIS decomposition plot
}
