% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_experiments.R
\name{all_combs_ensemble}
\alias{all_combs_ensemble}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
all_combs_ensemble(
  data,
  nmod = 3,
  window = 5,
  init_weeks = 5,
  avail_threshold = 0.5,
  avail_overlap_threshold = 0.1,
  excl = c("EuroCOVIDhub-baseline", "EuroCOVIDhub-ensemble")
)
}
\arguments{
\item{data}{data (subset or full) from the European Forecast hub}

\item{nmod}{number of models to sample at each date}

\item{window}{window size for recent distance computation}

\item{init_weeks}{number of weeks left out at beginning for init}

\item{avail_threshold}{minimum availability for models to be considered in the
sampling process}

\item{avail_overlap_threshold}{minimum common availability for model pair}

\item{excl}{which models should be excluded from the ensemble experiment}
}
\description{
Samples ALL combinations of possible ensembles of size n at given timepoint,
scores them and computes average pairwise historical and recent distance
}
