% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_experiments.R
\name{leaveout_ensemble}
\alias{leaveout_ensemble}
\title{COVID-19 Forecast Hub ensemble and model structure analysis}
\usage{
leaveout_ensemble(
  data,
  avail_threshold,
  nmods = c(5, 6, 7, 8, 100),
  samples = 100,
  seed = 32,
  excl = c("EuroCOVIDhub-baseline", "EuroCOVIDhub-ensemble")
)
}
\arguments{
\item{data}{data (subset or full) from the European Forecast hub}

\item{avail_threshold}{minimum availability for models to be considered in the
sampling process}

\item{nmods}{number of models that should be sampled for each step}

\item{samples}{how many random samples of models to draw}

\item{seed}{random seed for sampling}

\item{excl}{which models should be excluded from the ensemble experiment}
}
\description{
Assesses ensemble performance by iteratively sampling given numbers of
model for the (mean/median) ensemble
}
